node: expose `Encoding`
