+++ bindings/node/lib/index.ts
@@ -1,2 +1,2 @@
-// export * from './bindings';
+// export * from "./bindings";
 export * from "./tokenizers";
