+++ bindings/node/lib/tokenizers/index.ts
@@ -1,3 +1,4 @@
+export { Encoding } from "./base.tokenizer";
 export * from "./bert-wordpiece.tokenizer";
 export * from "./bpe.tokenizer";
 export * from "./byte-level-bpe.tokenizer";
